refactor: added component to the characteristics tabs
