+++ src/components/CharacterTraits.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import { ComputedRef, computed, ref, toRefs, watch } from 'vue';
+import { ComputedRef, computed, toRefs, watch } from 'vue';
 
+import CharacterTraitsTabs from '@components/CharacterTraitsTabs.vue';
 import { Episode } from '@domain/index';
 
 import { useColorTags } from '@/composables/useColorTags';
@@ -13,8 +14,6 @@ const { clearSelectedCharacter } = useSelectedCharacter();
 const { name, episode, gender, id, image, location, origin, species, status, type } = toRefs(store.selectedCharacter);
 
 const lastEpisode: ComputedRef<Episode> = computed(() => episode.value[episode.value.length - 1]);
-
-const tab = ref(null);
 
 const useTag = useColorTags(species.value, status.value || '');
 let statusColorOnTag = useTag.colorTagStatus;
@@ -69,96 +68,26 @@ watch(id, () => {
             <v-card-text>
                 <img class="character-traits__image" :src="image" alt="" />
             </v-card-text>
-
-            <v-card class="character-traits__sheet">
-                <v-tabs v-model="tab" class="character-traits__tabs" color="primary" align-tabs="center">
-                    <v-tab value="about">About</v-tab>
-                    <v-tab value="episodes">Episodes</v-tab>
-                    <v-tab value="clones">Clones</v-tab>
-                </v-tabs>
-
-                <v-card-text class="character-traits__sheet-content">
-                    <v-window v-model="tab">
-                        <v-window-item
-                            class="character-traits__tab-content"
-                            value="about"
-                            data-character-traits="about"
-                        >
-                            <v-list lines="five">
-                                <v-row no-gutters>
-                                    <v-col cols="5">
-                                        <v-list-item class="character-traits__item" title="Gender" :subtitle="gender">
-                                            <template #prepend>
-                                                <v-icon color="blue">mdi-gender-male-female</v-icon>
-                                            </template>
-                                        </v-list-item>
-                                    </v-col>
-                                    <v-col cols="7">
-                                        <v-list-item class="character-traits__item" title="Born in" :subtitle="origin">
-                                            <template #prepend>
-                                                <v-icon color="red">mdi-heart-pulse</v-icon>
-                                            </template>
-                                        </v-list-item>
-                                    </v-col>
-                                </v-row>
-                                <v-list-item class="character-traits__item" title="Type" :subtitle="type">
-                                    <template #prepend>
-                                        <v-icon color="purple">mdi-account-search</v-icon>
-                                    </template>
-                                </v-list-item>
-                                <v-list-item
-                                    class="character-traits__item"
-                                    title="Location"
-                                    data-character-traits="location"
-                                >
-                                    <template #prepend>
-                                        <v-icon color="green">mdi-earth-box</v-icon>
-                                    </template>
-
-                                    <template #subtitle>
-                                        <p>Name: {{ location.name }}</p>
-                                        <p>Dimension: {{ location.dimension }}</p>
-                                        <p>Type: {{ location.type }}</p>
-                                    </template>
-                                </v-list-item>
-                                <v-list-item
-                                    class="character-traits__item"
-                                    title="Last episode"
-                                    data-character-traits="last-episode"
-                                >
-                                    <template #prepend>
-                                        <v-icon color="black">mdi-video-box</v-icon>
-                                    </template>
-                                    <template #subtitle>
-                                        <p>Name: {{ lastEpisode.name }}</p>
-                                        <p>Season: {{ lastEpisode.season }}</p>
-                                        <p>Episode: {{ lastEpisode.number }}</p>
-                                    </template>
-                                </v-list-item>
-                            </v-list>
-                        </v-window-item>
-
-                        <v-window-item class="character-traits__tab-content" value="episodes">
-                            <v-chip v-for="(currentEpisode, index) in episode" :key="index" class="ma-2">
-                                S{{ currentEpisode.season }} - E{{ currentEpisode.season }}:
-                                {{ currentEpisode.name }} [{{ currentEpisode.air_date }}]
-                            </v-chip>
-                        </v-window-item>
-
-                        <v-window-item class="character-traits__tab-content" value="clones"> </v-window-item>
-                    </v-window>
-                </v-card-text>
-            </v-card>
+            <CharacterTraitsTabs
+                :gender="gender"
+                :origin="origin"
+                :type="type"
+                :location="location"
+                :last-episode="lastEpisode"
+                :episode="episode"
+            ></CharacterTraitsTabs>
         </v-card>
     </Teleport>
 </template>
 
 <style lang="postcss" scoped>
-body {
+body,
+#app {
     /* @apply overflow-hidden overflow-visible; */
 
-    scrollbar-width: none;
+    scrollbar-width: 0px;
     overflow: hidden;
+    overflow-y: hidden;
 
     &::-webkit-scrollbar {
         display: none;
@@ -199,7 +128,7 @@ body {
     }
 
     &__image {
-        @apply rounded-3xl sm:max-w-[250px] max-w-xs m-auto;
+        @apply rounded-3xl sm:max-w-[250px] tablet:max-w-xs m-auto;
     }
 
     &__sheet {
@@ -246,23 +175,6 @@ body {
 }
 </style>
 <style scoped>
-body {
-    /* @apply overflow-hidden overflow-visible; */
-
-    scrollbar-width: none !important;
-    overflow: hidden !important;
-
-    /* &::-webkit-scrollbar {
-        display: none;
-    } */
-}
-body::-webkit-scrollbar {
-    display: none !important;
-}
-.v-list-item__content > .v-list-item-subtitle {
-    text-transform: capitalize;
-}
-
 .character-traits .v-list-item__prepend > .v-icon {
     margin-right: 8px;
 }
